+++ R/utils.R
@@ -149,13 +149,6 @@ local_sqlite_connection <- function(envir = parent.frame()) {
   )
 }
 
-local_memdb_frame <- function(name, ..., frame = parent.frame()) {
-  df <- tibble::tibble(...)
-
-  withr::defer(DBI::dbRemoveTable(src_memdb()$con, name), envir = frame)
-  copy_to(src_memdb(), df, name, temporary = TRUE)
-}
-
 is_testing <- function() {
   identical(Sys.getenv("TESTTHAT"), "true")
 }