+++ tests/testthat/test-src_dbi.R
@@ -1,5 +1,5 @@
 test_that("tbl and src classes include connection class", {
-  mf <- memdb_frame(x = 1, y = 2)
+  mf <- local_memdb_frame("df", x = 1, y = 2)
   expect_true(inherits(mf, "tbl_SQLiteConnection"))
   expect_true(inherits(mf$src, "src_SQLiteConnection"))
 })