+++ tests/testthat/test-verb-collect.R
@@ -1,5 +1,5 @@
 test_that("collect equivalent to as.data.frame/as_tibble", {
-  mf <- memdb_frame(letters)
+  mf <- local_memdb_frame("df", letters)
 
   expect_equal(as.data.frame(mf), data.frame(letters, stringsAsFactors = FALSE))
   expect_equal(as_tibble(mf), tibble(letters))
@@ -8,7 +8,7 @@ test_that("collect equivalent to as.data.frame/as_tibble", {
 
 test_that("explicit collection returns all data", {
   n <- 1e5 + 10 # previous default was 1e5
-  big <- memdb_frame(x = seq_len(n))
+  big <- local_memdb_frame("df", x = seq_len(n))
 
   nrow1 <- big |> as.data.frame() |> nrow()
   nrow2 <- big |> as_tibble() |> nrow()
@@ -20,7 +20,7 @@ test_that("explicit collection returns all data", {
 })
 
 test_that("collect() handles DBI error", {
-  mf <- memdb_frame(x = 1)
+  mf <- local_memdb_frame("df", x = 1)
   expect_snapshot(
     (expect_error(mf |> mutate(a = sql("invalid sql")) |> collect())),
     transform = snap_transform_dbi