+++ tests/testthat/test-verb-uncount.R
@@ -30,7 +30,7 @@ test_that("symbols weights are dropped in output", {
 })
 
 test_that("can request to preserve symbols", {
-  df <- memdb_frame(x = 1, w = 1)
+  df <- local_memdb_frame("df", x = 1, w = 1)
 
   expect_equal(
     dbplyr_uncount(df, w, .remove = FALSE) |> colnames(),
@@ -39,7 +39,7 @@ test_that("can request to preserve symbols", {
 })
 
 test_that("unique identifiers created on request", {
-  df <- memdb_frame(w = 1:3)
+  df <- local_memdb_frame("df", w = 1:3)
   expect_equal(
     dbplyr_uncount(df, w, .id = "id") |> collect() |> arrange(id),
     tibble(id = c(1L, 1:2, 1:3)) |> arrange(id)
@@ -47,7 +47,7 @@ test_that("unique identifiers created on request", {
 })
 
 test_that("expands constants and expressions", {
-  df <- memdb_frame(x = 1, w = 2)
+  df <- local_memdb_frame("df", x = 1, w = 2)
 
   expect_equal(dbplyr_uncount(df, 2) |> collect(), collect(df)[c(1, 1), ])
   expect_equal(dbplyr_uncount(df, 1 + 1) |> collect(), collect(df)[c(1, 1), ])
@@ -55,24 +55,24 @@ test_that("expands constants and expressions", {
 
 
 test_that("works with groups", {
-  df <- memdb_frame(g = 1, x = 1, w = 1) |> dplyr::group_by(g)
+  df <- local_memdb_frame("df", g = 1, x = 1, w = 1) |> dplyr::group_by(g)
   expect_equal(group_vars(dbplyr_uncount(df, w)), "g")
 })
 
 test_that("grouping variable are removed", {
-  df <- memdb_frame(g = 1, x = 1, w = 1) |> dplyr::group_by(g)
+  df <- local_memdb_frame("df", g = 1, x = 1, w = 1) |> dplyr::group_by(g)
 
   expect_equal(dbplyr_uncount(df, g) |> colnames(), c("x", "w"))
 })
 
 test_that("must evaluate to integer", {
-  df <- memdb_frame(x = 1, w = 1 / 2)
+  df <- local_memdb_frame("df", x = 1, w = 1 / 2)
   expect_error(dbplyr_uncount(df, w), class = "vctrs_error_cast_lossy")
 
   expect_error(dbplyr_uncount(df, "W"), class = "vctrs_error_incompatible_type")
 })
 
 test_that("works with 0 weights", {
-  df <- memdb_frame(x = 1:2, w = c(0, 1))
+  df <- local_memdb_frame("df", x = 1:2, w = c(0, 1))
   expect_equal(dbplyr_uncount(df, w) |> collect(), tibble(x = 2))
 })